requirements.txt
torch>=1.9.0
torchvision>=0.10.0
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.19.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
albumentations>=1.0.0
Pillow>=8.0.0
shapely>=1.7.0
requests>=2.25.0
tqdm>=4.60.0
supervision>=0.26.0
fastapi>=0.110.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0

black>=24.0.0
flake8>=7.0.0
pytest>=8.0.0

pydantic-settings~=2.13.1